Minor tweaks
Pointed the GitHub link to the site's GitHub repo (seems smart to have a tie back to the code, and my personal GH profile is not particular well groomed ATM anyway). For the GitHub repo, added a readme file. In both the readme and the log file, attempted relative links; may have to give them another go later.